a replacement module for plumed when not available
