+++ src/LangevinIntegrators.jl
@@ -53,32 +53,11 @@ export addFix!
 try
     include("plumed.jl") # Find a way to make this optionnal, using Require?
 catch e  # Allow to proceed even when plumed is not found
-    warn("Unable to load plumed extension: ",e)
-    warn("A empty plumed module have been loaded instead, it has not effect.")
+    @warn("Unable to load plumed extension: ",e)
+    @warn("A empty plumed module have been loaded instead, it has not effect.")
 
-    module Plumed
-        struct plumed{TF<: AbstractFloat} <: AbstractFix
-                function plumed(plumed_input_file::String, plumed_log_file::String="p.log", dim::Int64=1, delta_t::Float64=1e-3; temperature=1.0)
-                    return new()
-                end
-        end
-        function init_fix(fix::plumed; kwargs...) # Différent du constructeur au sens que c'est appelé au début de chaque traj
-            warn("Plumed has not been loaded, this is an empty fix.")
-            return fix
-        end
+    include("plumed_replacement.jl")
 
-        function apply_fix!(fix::plumed,x::Array{TF},f::Array{TF}; kwargs...) where {TF<:AbstractFloat}
-        	return false
-        end
-
-
-        function close_fix(fix::plumed)
-        end
-
-        function addPlumed!(integrator::AbstractIntegrator,plumed_input_file::String, plumed_log_file::String="p.log", dim::Int64=1; temperature=1.0)
-            addFix!(integrator, plumed(plumed_input_file, plumed_log_file, dim, integrator.Δt; temperature=temperature))
-        end
-    end
 end
 
 include("generate_initcond.jl")
